+++ changelog.md
@@ -2,6 +2,9 @@
 ## [0.24.0] - UNRELEASED
 ### Breaking Changes
 * Database readers and writes for each `BaseService` no longer accept a `MetadataCollection`, and will instead use the collection of the provided service.
+* Tracing downstream/upstream using a NetworkTrace will not step on a start item if there is no downstream/upstream terminal on the item.
+* `SetDirection` now correctly applies te `BOTH` direction on all parts of the loop again, so if you were relying on the broken intermediate state, you will
+  need to update your code.
 
 ### New Features
 * Network state services for updating and querying network state events via gRPC.
@@ -22,9 +25,6 @@
 
 ## [0.23.0] - 2024-10-18
 ### Breaking Changes
-* Tracing downstream/upstream using a NetworkTrace will not step on a start item if there is no downstream/upstream terminal on the item.
-
-
 * Updated to latest evolve-grpc major version.
 * Removed unused AuthType enum.
 * Removed unused kotlinx-serialization-json dependency.